+++ praw/models/__init__.py
@@ -11,6 +11,8 @@ from .listing.domain import DomainListing
 from .listing.generator import ListingGenerator
 from .listing.listing import Listing, ModeratorListing, ModmailConversationsListing
 from .mod_action import ModAction
+from .mod_note import ModNote
+from .mod_notes import RedditModNotes, RedditorModNotes, SubredditModNotes
 from .preferences import Preferences
 from .reddit.collections import Collection
 from .reddit.comment import Comment