+++ tests/unit/models/listing/test_listing.py
@@ -1,4 +1,4 @@
-from praw.models.listing.listing import ModmailConversationsListing
+from praw.models.listing.listing import ModmailConversationsListing, ModNoteListing
 
 from ... import UnitTest
 
@@ -8,4 +8,14 @@ class TestModmailConversationsListing(UnitTest):
         assert (
             ModmailConversationsListing(self.reddit, _data={"conversations": []}).after
             is None
+        )
+
+
+class TestModNoteListing(UnitTest):
+    def test_has_next_page(self):
+        assert (
+            ModNoteListing(
+                self.reddit, _data={"has_next_page": True, "end_cursor": "end_cursor"}
+            ).after
+            == "end_cursor"
         )