+++ tests/unit/models/reddit/test_subreddit.py
@@ -95,6 +95,14 @@ class TestSubreddit(UnitTest):
         with pytest.raises(ServerError):
             self.reddit.subreddit("test").submit_image("Test", "/dev/null")
 
+    def test_notes_delete__invalid_args(self):
+        with pytest.raises(TypeError) as excinfo:
+            self.reddit.subreddit("SubTestBot1").mod.notes.delete(note_id="111")
+        assert excinfo.value.args[0] == (
+            "Either the `redditor` parameter must be provided or this method must be"
+            " called from a Redditor instance (e.g., `redditor.notes`)."
+        )
+
     def test_pickle(self):
         subreddit = Subreddit(
             self.reddit, _data={"display_name": "name", "id": "dummy"}